+++ blablatwoApp/src/main/java/com/blablatwo/city/CityRepository.java
@@ -4,13 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CityRepository {
-    Optional<CityEntity> findById(int id);
+    Optional<CityEntity> findById(long id);
 
     List<CityEntity> findAll();
 
     CityEntity save(CityEntity city);
 
-    boolean deleteById(int id);
+    boolean deleteById(long id);
 
     Optional<CityEntity> update(CityEntity city);
 }
